global state implemented
