+++ contexts/state.js
@@ -1,0 +1,64 @@
+import axios from "axios";
+import { useContext, useEffect, useState } from "react";
+import { createContext } from "react";
+
+export const GlobalContext = createContext({});
+
+export default function GlobalProvider({ children }) {
+  // const [pokemonDetail, setPokemonDetail] = useState({});
+  const [pokelist, setPokelist] = useState([]);
+
+  function getPokemonColor(pokemonType) {
+    switch (pokemonType) {
+      case "bug":
+        return "#316520";
+      case "dark":
+        return "#5C5365";
+      case "dragon":
+        return "#0A6CBF";
+      case "electric":
+        return "#F4D23B";
+      case "fairy":
+        return "#EC8FE6";
+      case "fighting":
+        return "#CE4069";
+      case "fire":
+        return "#F44900";
+      case "flying":
+        return "#6892B0";
+      case "ghost":
+        return "#5269AC";
+      case "grass":
+        return "#70B873";
+      case "ground":
+        return "#D97745";
+      case "ice":
+        return "#74CEC0";
+      case "normal":
+        return "#8A8A8A";
+      case "poison":
+        return "#AD61AE";
+      case "psychic":
+        return "#F67176";
+      case "rock":
+        return "#C7B78B";
+      case "steel":
+        return "#BBBBBB";
+      case "water":
+        return "#33A4F5";
+      default:
+        return "#8A8A8A";
+    }
+  }
+
+
+  return (
+    <GlobalContext.Provider value={{ 
+      pokelist,
+      setPokelist,
+      getPokemonColor
+    }}>
+      {children}
+    </GlobalContext.Provider>
+  );
+}
