+++ pages/_app.js
@@ -1,10 +1,12 @@
 import GlobalStyles from "../globalStyles";
+import GlobalProvider from "../contexts/state";
 
 export default function MyApp({ Component, pageProps }) {
+
   return (
-    <>
+    <GlobalProvider>
       <GlobalStyles />
       <Component {...pageProps} />
-    </>
+    </GlobalProvider>
   );
 }
