+++ src/components/PokemonCard/index.js
@@ -1,10 +1,13 @@
-import axios from "axios";
-import { useEffect, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import { CardColor, PokemonCardStyled } from "./styles";
 import { getPokemonType } from "./types";
 import Link from "next/link";
+import axios from "axios";
+import { GlobalContext } from "../../../contexts/state";
 
 export default function PokemonCard({ pokemon }) {
+  const { getPokemonColor } = useContext(GlobalContext);
+
   const [pokemonDetail, setPokemonDetail] = useState({});
 
   const fetchPokemonDetail = async () => {
@@ -20,48 +23,9 @@ export default function PokemonCard({ pokemon }) {
     fetchPokemonDetail();
   }, []);
 
-  function getPokemonColor(pokemonType) {
-    switch (pokemonType) {
-      case "bug":
-        return "#316520";
-      case "dark":
-        return "#5C5365";
-      case "dragon":
-        return "#0A6CBF";
-      case "electric":
-        return "#F4D23B";
-      case "fairy":
-        return "#EC8FE6";
-      case "fighting":
-        return "#CE4069";
-      case "fire":
-        return "#F44900";
-      case "flying":
-        return "#6892B0";
-      case "ghost":
-        return "#5269AC";
-      case "grass":
-        return "#70B873";
-      case "ground":
-        return "#D97745";
-      case "ice":
-        return "#74CEC0";
-      case "normal":
-        return "#8A8A8A";
-      case "poison":
-        return "#AD61AE";
-      case "psychic":
-        return "#F67176";
-      case "rock":
-        return "#C7B78B";
-      case "steel":
-        return "#BBBBBB";
-      case "water":
-        return "#33A4F5";
-      default:
-        return "#8A8A8A";
-    }
-  }
+
+
+
 
   return (
     <PokemonCardStyled>
@@ -89,7 +53,7 @@ export default function PokemonCard({ pokemon }) {
         </div>
         <hr />
         <div className="container-buttons-pokemon">
-          <Link href="/details/[id]" as={`/details/${pokemon.name}`}>
+          <Link href="/details/[id]" as={`/details/${pokemonDetail.id}`}>
             <button>Details</button>
           </Link>
           <button>Catch!</button>
